IntervalLinearAlgebra

Documentation for IntervalLinearAlgebra.

IntervalLinearAlgebra.AbstractDirectSolver

IntervalLinearAlgebra.AbstractIterativeSolver

IntervalLinearAlgebra.AbstractLinearSolver

IntervalLinearAlgebra.AbstractPrecondition

IntervalLinearAlgebra.GaussSeidel

IntervalLinearAlgebra.GaussianElimination

IntervalLinearAlgebra.HansenBliekRohn

IntervalLinearAlgebra.InverseDiagonalMidpoint

IntervalLinearAlgebra.InverseMidpoint

IntervalLinearAlgebra.Jacobi

IntervalLinearAlgebra.LinearKrawczyk

IntervalLinearAlgebra.NoPrecondition

IntervalLinearAlgebra.NonLinearOettliPrager

IntervalLinearAlgebra.comparison_matrix

IntervalLinearAlgebra.enclose

IntervalLinearAlgebra.interval_isapprox

IntervalLinearAlgebra.interval_norm

IntervalLinearAlgebra.interval_norm

IntervalLinearAlgebra.is_H_matrix

IntervalLinearAlgebra.is_M_matrix

IntervalLinearAlgebra.is_Z_matrix

IntervalLinearAlgebra.is_strictly_diagonally_dominant

IntervalLinearAlgebra.is_strongly_regular

IntervalLinearAlgebra.oettli_eq

IntervalLinearAlgebra.oettli_lhs

IntervalLinearAlgebra.oettli_rhs

IntervalLinearAlgebra.rref

IntervalLinearAlgebra.rref!

IntervalLinearAlgebra.AbstractDirectSolver — Type

AbstractDirectSolver <: AbstractLinearSolver

Abstract type for direct solvers of interval linear systems, such as Gaussian elimination and Hansen-Bliek-Rohn.

IntervalLinearAlgebra.AbstractIterativeSolver — Type

AbstractIterativeSolver <: AbstractLinearSolver

Abstract type for iterative solvers of interval linear systems, such as Jacobi or Gauss-Seidel.

IntervalLinearAlgebra.AbstractLinearSolver — Type

AbstractLinearSolver

Abstract type for solvers of interval linear systems.

IntervalLinearAlgebra.AbstractPrecondition — Type

AbstractPrecondition

Abstract type for preconditioners of interval linear systems.

IntervalLinearAlgebra.GaussSeidel — Type

GaussSeidel <: AbstractIterativeSolver

Type for the Gauss-Seidel solver of the interval linear system $Ax=b$. For details see Section 5.7.4 of [HOR19]

Fields

max_iterations – maximum number of iterations (default 20)

atol – absolute tolerance (default 0), if at some point $|xₖ - xₖ₊₁| < atol$ (elementwise), then stop and return $xₖ₊₁$. If atol=0, then min(diam(A))*1e-5 is used.

Notes

An object of type GaussSeidel is a function with method

  (gs::GaussSeidel)(A::AbstractMatrix{T},
                    b::AbstractVector{T},
                    [x]::AbstractVector{T}=enclose(A, b)) where {T<:Interval}

Input

A – N×N interval matrix

b – interval vector of length N

x – (optional) initial enclosure for the solution of $Ax = b$. If not given, it is automatically computed using enclose

Examples

julia> A = [2..4 -1..1;-1..1 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-1, 1]
 [-1, 1]   [2, 4]

julia> b = [-2..2, -1..1]
2-element Vector{Interval{Float64}}:
 [-2, 2]
 [-1, 1]

julia> gs = GaussSeidel()
GaussSeidel linear solver
max_iterations = 20
atol = 0.0

julia> gs(A, b)
2-element Vector{Interval{Float64}}:
 [-1.66668, 1.66668]
 [-1.33334, 1.33334]

IntervalLinearAlgebra.GaussianElimination — Type

GaussianElimination <: AbstractDirectSolver

Type for the Gaussian elimination solver of the square interval linear system $Ax=b$. For more details see section 5.6.1 of [HOR19]

Notes

An object of type GaussianElimination is a callable function with method

  (ge::GaussianElimination)(A::AbstractMatrix{T},
                            b::AbstractVector{T}) where {T<:Interval}

Examples

julia> A = [2..4 -1..1;-1..1 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-1, 1]
 [-1, 1]   [2, 4]

julia> b = [-2..2, -1..1]
2-element Vector{Interval{Float64}}:
 [-2, 2]
 [-1, 1]

julia> ge = GaussianElimination()
GaussianElimination linear solver

julia> ge(A, b)
2-element Vector{Interval{Float64}}:
 [-1.66667, 1.66667]
 [-1.33334, 1.33334]

IntervalLinearAlgebra.HansenBliekRohn — Type

HansenBliekRohn <: AbstractDirectSolver

Type for the HansenBliekRohn solver of the square interval linear system $Ax=b$. For more details see section 5.6.2 of [HOR19]

Notes

Hansen-Bliek-Rohn works with H-matrices without precondition and with strongly regular matrices using InverseMidpoint precondition

If the midpoint of $A$ is a diagonal matrix, then the algorithm returns the exact hull.

An object of type Hansen-Bliek-Rohn is a callable function with method

  (hbr::HansenBliekRohn)(A::AbstractMatrix{T},
                         b::AbstractVector{T}) where {T<:Interval}

Examples

julia> A = [2..4 -1..1;-1..1 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-1, 1]
 [-1, 1]   [2, 4]

julia> b = [-2..2, -1..1]
2-element Vector{Interval{Float64}}:
 [-2, 2]
 [-1, 1]

julia> hbr = HansenBliekRohn()
HansenBliekRohn linear solver

julia> hbr(A, b)
2-element Vector{Interval{Float64}}:
 [-1.66667, 1.66667]
 [-1.33334, 1.33334]

IntervalLinearAlgebra.InverseDiagonalMidpoint — Type

InverseDiagonalMidpoint <: Precondition

Preconditioner that preconditions the linear system $Ax=b$ with the diagonal matrix of $A_c^{-1}$, where $A_c$ is the midpoint matrix of $A$.

Notes

An object of type InverseDiagonalMidpoint is a function with method

  (idmp::InverseDiagonalMidpoint)(A::AbstractMatrix{T},
                                  b::AbstractVector{T}) where {T<:Interval}

Example

julia> A = [2..4 -2..1; -1..2 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-2, 1]
 [-1, 2]   [2, 4]

julia> b = [-2..2, -2..2]
2-element Vector{Interval{Float64}}:
 [-2, 2]
 [-2, 2]

julia> idmp = InverseDiagonalMidpoint()
InverseDiagonalMidpoint()

julia> idmp(A, b)
(Interval{Float64}[[0.666666, 1.33334] [-0.666667, 0.333334]; [-0.333334, 0.666667] [0.666666, 1.33334]], Interval{Float64}[[-0.666667, 0.666667], [-0.666667, 0.666667]])

IntervalLinearAlgebra.InverseMidpoint — Type

InverseMidpoint <: Precondition

Preconditioner that preconditions the linear system $Ax=b$ with $A_c^{-1}$, where $A_c$ is the midpoint matrix of $A$.

Notes

An object of type NoPrecondition is a function with method

  (imp::InverseMidpoint)(A::AbstractMatrix{T},
                         b::AbstractVector{T}) where {T<:Interval}

Examples

julia> A = [2..4 -2..1; -1..2 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-2, 1]
 [-1, 2]   [2, 4]

julia> b = [-2..2, -2..2]
2-element Vector{Interval{Float64}}:
 [-2, 2]
 [-2, 2]

julia> imp = InverseMidpoint()
InverseMidpoint()

julia> imp(A, b)
(Interval{Float64}[[0.594594, 1.40541] [-0.540541, 0.540541]; [-0.540541, 0.540541] [0.594594, 1.40541]], Interval{Float64}[[-0.756757, 0.756757], [-0.756757, 0.756757]])

IntervalLinearAlgebra.Jacobi — Type

Jacobi <: AbstractIterativeSolver

Type for the Jacobi solver of the interval linear system $Ax=b$. For details see Section 5.7.4 of [HOR19]

Fields

max_iterations – maximum number of iterations (default 20)

atol – absolute tolerance (default 0), if at some point $|xₖ - xₖ₊₁| < atol$ (elementwise), then stop and return $xₖ₊₁$. If atol=0, then min(diam(A))*1e-5 is used.

Notes

An object of type Jacobi is a function with method

  (jac::Jacobi)(A::AbstractMatrix{T},
                b::AbstractVector{T},
                [x]::AbstractVector{T}=enclose(A, b)) where {T<:Interval}

Input

A – N×N interval matrix

b – interval vector of length N

x – (optional) initial enclosure for the solution of $Ax = b$. If not given, it is automatically computed using enclose

Examples

julia> A = [2..4 -1..1;-1..1 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-1, 1]
 [-1, 1]   [2, 4]

julia> b = [-2..2, -1..1]
2-element Vector{Interval{Float64}}:
 [-2, 2]
 [-1, 1]

julia> jac = Jacobi()
Jacobi linear solver
max_iterations = 20
atol = 0.0

julia> jac(A, b)
2-element Vector{Interval{Float64}}:
 [-1.66668, 1.66668]
 [-1.33335, 1.33335]

IntervalLinearAlgebra.LinearKrawczyk — Type

LinearKrawczyk <: AbstractIterativeSolver

Type for the Krawczyk solver of the interval linear system $Ax=b$. For details see Section 5.7.3 of [HOR19]

Fields

max_iterations – maximum number of iterations (default 20)

atol – absolute tolerance (default 0), if at some point $|xₖ - xₖ₊₁| < atol$ (elementwise), then stop and return $xₖ₊₁$. If atol=0, then min(diam(A))*1e-5 is used.

Notes

An object of type LinearKrawczyk is a function with method

  (kra::LinearKrawczyk)(A::AbstractMatrix{T},
                        b::AbstractVector{T},
                        [x]::AbstractVector{T}=enclose(A, b)) where {T<:Interval}

Input

A – N×N interval matrix

b – interval vector of length N

x – (optional) initial enclosure for the solution of $Ax = b$. If not given, it is automatically computed using enclose

Examples

julia> A = [2..4 -1..1;-1..1 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-1, 1]
 [-1, 1]   [2, 4]

julia> b = [-2..2, -1..1]
2-element Vector{Interval{Float64}}:
 [-2, 2]
 [-1, 1]

julia> kra = LinearKrawczyk()
LinearKrawczyk linear solver
max_iterations = 20
atol = 0.0

julia> kra(A, b)
2-element Vector{Interval{Float64}}:
 [-2, 2]
 [-2, 2]

IntervalLinearAlgebra.NoPrecondition — Type

NoPrecondition <: AbstractPrecondition

Type of the trivial preconditioner which does nothing.

Notes

An object of type NoPrecondition is a function with method

  (np::NoPrecondition)(A::AbstractMatrix{T},
                       b::AbstractVector{T}) where {T<:Interval}

Example

julia> A = [2..4 -2..1; -1..2 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-2, 1]
 [-1, 2]   [2, 4]

julia> b = [-2..2, -2..2]
2-element Vector{Interval{Float64}}:
 [-2, 2]
 [-2, 2]

julia> np = NoPrecondition()
NoPrecondition()

julia> np(A, b)
(Interval{Float64}[[2, 4] [-2, 1]; [-1, 2] [2, 4]], Interval{Float64}[[-2, 2], [-2, 2]])

IntervalLinearAlgebra.NonLinearOettliPrager — Type

NonLinearOettliPrager <: AbstractLinearSolver

Type for the OettliPrager solver of the interval linear system $Ax=b$. The solver first converts the system of interval equalities into a system of real inequalities using Oettli-Präger theorem [OET64] and then finds the feasible set using the forward-backward contractor method [JAU14] implemented in IntervalConstraintProgramming.jl.

Fields

tol – tolerance for the paving, default 0.01.

Notes

You need to import IntervalConstraintProgramming.jl to use this functionality.

An object of type OettliPrager is a function with methods

  (op::OettliPrager)(A::AbstractMatrix{T},
                     b::AbstractVector{T},
                     [X]::AbstractVector{T}=enclose(A, b)) where {T<:Interval}

  (op::OettliPrager)(A::AbstractMatrix{T},
                     b::AbstractVector{T},
                     X::IntervalBox) where {T<:Interval}

Input

A – N×N interval matrix

b – interval vector of length N

X – (optional) initial enclosure for the solution of $Ax = b$. If not given, it is automatically computed using enclose

Examples

julia> A = [2..4 -2..1;-1..2 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-2, 1]
 [-1, 2]   [2, 4]

julia> b = [-2..2, -2..2]
2-element Vector{Interval{Float64}}:
 [-2, 2]
 [-2, 2]

julia> op = NonLinearOettliPrager(0.1)
NonLinearOettliPrager linear solver
tol = 0.1

julia> op(A, b)
Paving:
- tolerance ϵ = 0.1
- inner approx. of length 1195
- boundary approx. of length 823

IntervalLinearAlgebra.comparison_matrix — Method

comparison_matrix(A::AbstractMatrix{T}) where {T<:Interval}

Computes the comparison matrix $⟨A⟩$ of the given interval matrix $A$ according to the definition $⟨A⟩ᵢᵢ = mig(Aᵢᵢ)$ and $⟨A⟩ᵢⱼ = -mag(Aᵢⱼ)$ if $i≠j$.

Examples

julia> A = [2..4 -1..1; -1..1 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-1, 1]
 [-1, 1]   [2, 4]

julia> comparison_matrix(A)
2×2 Matrix{Float64}:
  2.0  -1.0
 -1.0   2.0

IntervalLinearAlgebra.enclose — Method

enclose(A::AbstractMatrix{T}, b::AbstractVector{T}) where {T<:Interval}

Computes an enclosure of the solution of the interval linear system $Ax=b$ using the algorithm described in sec. 5.7.1 of [HOR19].

IntervalLinearAlgebra.interval_isapprox — Method

interval_isapprox(a::Interval, b::Interval; kwargs)

Checks whether the intervals $a$ and $b$ are approximate equal, that is both their lower and upper bound are approximately equal.

Keywords

Same of Base.isapprox

Example

julia> a = 1..2
[1, 2]

julia> b = a + 1e-10
[1, 2.00001]

julia> interval_isapprox(a, b)
true

julia> interval_isapprox(a, b; atol=1e-15)
false

IntervalLinearAlgebra.interval_norm — Method

interval_norm(A::AbstractMatrix{T}) where {T<:Interval}

computes the infinity norm of interval matrix $A$.

Examples

julia> A = [2..4 -1..1; -1..1 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-1, 1]
 [-1, 1]   [2, 4]

julia> interval_norm(A)
5.0

IntervalLinearAlgebra.interval_norm — Method

interval_norm(A::AbstractVector{T}) where {T<:Interval}

computes the infinity norm of interval vector $v$.

Examples

julia> b = [-2..2, -3..2]
2-element Vector{Interval{Float64}}:
 [-2, 2]
 [-3, 2]

julia> interval_norm(b)
3.0

IntervalLinearAlgebra.is_H_matrix — Method

is_H_matrix(A::AbstractMatrix{T}) where {T<:Interval}

Tests whether the square interval matrix A is an H-matrix, by testing that $⟨A⟩^{-1}e>0$, where $e=[1, 1, …, 1]ᵀ$. Note that in practice it tests that a floating point approximation of $⟨A⟩^{-1}e$ satisfies the condition. For more details see section 4.4 of [HOR19].

Examples

julia> A = [2..4 -1..1; -1..1 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-1, 1]
 [-1, 1]   [2, 4]

julia> is_H_matrix(A)
true

julia> A = [2..4 -2..1; -1..2 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-2, 1]
 [-1, 2]   [2, 4]

julia> is_H_matrix(A)
false

IntervalLinearAlgebra.is_M_matrix — Method

is_M_matrix(A::AbstractMatrix{T}) where {T<:Interval}

Checks whether the square interval matrix $A$ is an M-matrix, that is a Z-matrix with non-negative inverse. For more details see section 4.2 of [HOR19].

Examples

julia> A = [2..2 -1..0; -1..0 2..2]
2×2 Matrix{Interval{Float64}}:
  [2, 2]  [-1, 0]
 [-1, 0]   [2, 2]

julia> is_M_matrix(A)
true

julia> A = [2..4 -2..1; -1..2 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-2, 1]
 [-1, 2]   [2, 4]

julia> is_M_matrix(A)
false

IntervalLinearAlgebra.is_Z_matrix — Method

is_Z_matrix(A::AbstractMatrix{T}) where {T<:Interval}

Checks whether the square interval matrix $A$ is a Z-matrix, that is whether $Aᵢⱼ≤0$ for all $i≠j$. For more details see section 4.2 of [HOR19].

Examples

julia> A = [2..4 -2.. -1; -2.. -1 2..4]
2×2 Matrix{Interval{Float64}}:
   [2, 4]  [-2, -1]
 [-2, -1]    [2, 4]

julia> is_Z_matrix(A)
true

julia> A = [2..4 -2..1; -1..2 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-2, 1]
 [-1, 2]   [2, 4]

julia> is_Z_matrix(A)
false

IntervalLinearAlgebra.is_strictly_diagonally_dominant — Method

is_strictly_diagonally_dominant(A::AbstractMatrix{T}) where {T<:Interval}

Checks whether the square interval matrix $A$ of order $n$ is stictly diagonally dominant, that is if $mig(Aᵢᵢ) > ∑_{k ≠ i} mag(Aᵢₖ)$ for $i=1,…,n$. For more details see section 4.5 of [HOR19].

Examples

julia> A = [2..4 -1..1; -1..1 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-1, 1]
 [-1, 1]   [2, 4]

julia> is_strictly_diagonally_dominant(A)
true

julia> A = [2..4 -2..1; -1..2 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-2, 1]
 [-1, 2]   [2, 4]

julia> is_strictly_diagonally_dominant(A)
false

IntervalLinearAlgebra.is_strongly_regular — Method

is_strongly_regular(A::AbstractMatrix{T}) where {T<:Interval}

Tests whether the square interval matrix $A$ is strongly regular, i.e. if $A_c^{-1}A$ is an H-matrix, where $A_c$ is the midpoint matrix of $A$`. For more details see section 4.6 of [HOR19].

Examples

julia> A = [2..4 -2..1; -1..2 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-2, 1]
 [-1, 2]   [2, 4]

julia> is_strongly_regular(A)
true

julia> A = [0..2 1..1;-1.. -1 0..2]
2×2 Matrix{Interval{Float64}}:
   [0, 2]  [1, 1]
 [-1, -1]  [0, 2]

julia> is_strongly_regular(A)
false

IntervalLinearAlgebra.oettli_eq — Method

Returns the separator for the constraint |a_c ⋅x - b_c| - a_r ⋅|x| - b_r <= 0.

a and x must be vectors of the same length and b is scalar.

The absolue values in the equation are taken elementwise.

a_c and a_r are vectors containing midpoints and radii of the intervals in a. Similar b_c and b_r.

IntervalLinearAlgebra.oettli_lhs — Method

returns the unrolled expression for $|a ⋅x - b|$ $a$ and $x$ must be vectors of the same length and $b$ is scalar. The absolue value in the equation is taken elementwise.

IntervalLinearAlgebra.oettli_rhs — Method

returns the unrolled expression for $a ⋅|x| + b$ $a$ and $x$ must be vectors of the same length and $b$ is scalar. The absolue value in the equation is taken elementwise.

IntervalLinearAlgebra.rref! — Method

rref!(A::AbstractMatrix{T}) where {T<:Interval}

In-place version of rref.

IntervalLinearAlgebra.rref — Method

rref(A::AbstractMatrix{T}) where {T<:Interval}

Computes the reduced row echelon form of the interval matrix A using maximum mignitude as pivoting strategy.

Examples

julia> A = [2..4 -1..1; -1..1 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-1, 1]
 [-1, 1]   [2, 4]

julia> rref(A)
2×2 Matrix{Interval{Float64}}:
 [2, 4]  [-1, 1]
 [0, 0]       [1.5, 4.5]