IntervalLinearAlgebra

Documentation for IntervalLinearAlgebra.

IntervalLinearAlgebra.GaussElimination

IntervalLinearAlgebra.GaussSeidel

IntervalLinearAlgebra.HansenBliekRohn

IntervalLinearAlgebra.InverseDiagonalMidpoint

IntervalLinearAlgebra.InverseMidpoint

IntervalLinearAlgebra.Jacobi

IntervalLinearAlgebra.Krawczyk

IntervalLinearAlgebra.NoPrecondition

IntervalLinearAlgebra.OettliPrager

IntervalLinearAlgebra.comparison_matrix

IntervalLinearAlgebra.enclose

IntervalLinearAlgebra.interval_norm

IntervalLinearAlgebra.is_H_matrix

IntervalLinearAlgebra.is_M_matrix

IntervalLinearAlgebra.is_Z_matrix

IntervalLinearAlgebra.is_strictly_diagonally_dominant

IntervalLinearAlgebra.is_strongly_regular

IntervalLinearAlgebra.oettli_eq

IntervalLinearAlgebra.oettli_lhs

IntervalLinearAlgebra.oettli_rhs

IntervalLinearAlgebra.rref

IntervalLinearAlgebra.GaussElimination — Type

GaussElimination <: LinearSolver

Returns a Gaussian elimination solver for the interval linear system Ax=b.

IntervalLinearAlgebra.GaussSeidel — Type

GaussSeidel(max_iterations, atol)

Returns a Gauss-Seidel solver for the interval linear system Ax=b.

PARAMETERS:

max_iterations: maximum number of iterations (default 20)

atol: absolute tolerance (default 0), if at some point |xₖ - xₖ₊₁| < atol (elementwise), then stop and return xₖ₊₁. If atol=0, then min(diam(A))*1e-5 is used.

IntervalLinearAlgebra.HansenBliekRohn — Type

HansenBliekRohn()

Returns a Hansen-Bliek-Rohn solver for the interval linear system Ax=b.

IntervalLinearAlgebra.InverseDiagonalMidpoint — Type

InverseDiagonalMidpoint <: Precondition

Preconditioner that preconditions the linear system $Ax=b$ with $inv(Diagonal(Ac))$, where Ac is the midpoint matrix of $A$.

Example

idmp = InverseDiagonalMidpoint() # instantiate preconditioner
Aprec, bprec = idmp(A, b) # apply preconditioner

IntervalLinearAlgebra.InverseMidpoint — Type

InverseMidpoint <: Precondition

Preconditioner that preconditions the linear system $Ax=b$ with $inv(Ac)$, where Ac is the midpoint matrix of $A$.

Example

imp = InverseMidpoint() # instantiate preconditioner
Aprec, bprec = imp(A, b) # apply preconditioner

IntervalLinearAlgebra.Jacobi — Type

Jacobi(max_iterations, atol)

Returns a Jacobi solver for the interval linear system Ax=b.

PARAMETERS:

max_iterations: maximum number of iterations (default 20)

atol: absolute tolerance (default 0), if at some point |xₖ - xₖ₊₁| < atol (elementwise), then stop and return xₖ₊₁. If atol=0, then min(diam(A))*1e-5 is used.

IntervalLinearAlgebra.Krawczyk — Type

Krawczyk(max_iterations, atol)

Returns a Krawczyk solver for the interval linear system Ax=b.

PARAMETERS:

max_iterations: maximum number of iterations (default 20)

atol: absolute tolerance (default 0), if at some point |xₖ - xₖ₊₁| < atol (elementwise), then stop and return xₖ₊₁. If atol=0, then min(diam(A))*1e-5 is used.

IntervalLinearAlgebra.NoPrecondition — Type

NoPrecondition <: Precondition

Trivial preconditioner which does nothing.

Example

np = NoPrecondition() # instantiate preconditioner
Aprec, bprec = np(A, b) # apply preconditioner

IntervalLinearAlgebra.OettliPrager — Type

OettliPrager <: LinearSolver

Type for the OettliPrager solver of the interval linear system $Ax=b$. The solver first converts the system of interval equalities into a system of real inequalities using Oettli-Präger theorem [OET64] and then finds the feasible set using the forward-backward contractor method [JAU14] implemented in IntervalConstraintProgramming.jl.

Fields

tol – tolerance for the paving, default 0.01.

Notes

You need to import IntervalConstraintProgramming.jl to use this functionality.

An object of type OettliPrager is a function with methods

  (op::OettliPrager)(A::AbstractMatrix{T},
                  b::AbstractVector{T},
                  [X]::AbstractVector{T}=enclose(A, b)) where {T<:Interval}

  (op::OettliPrager)(A::AbstractMatrix{T},
                  b::AbstractVector{T},
                  X::IntervalBox) where {T<:Interval}

Input

A – N×N interval matrix

b – interval vector of length N

X – (optional) initial enclosure for the solution of $Ax = b$. If not given, it is automatically computed using enclose

Examples

julia> A = [2..4 -2..1; -1..2 2..4]
2×2 Matrix{Interval{Float64}}:
  [2, 4]  [-2, 1]
 [-1, 2]   [2, 4]

julia> b = [-2..2, -2..2]
2-element Vector{Interval{Float64}}:
 [-2, 2]
 [-2, 2]

julia> op = OettliPrager(0.1)
OettliPrager linear solver
tol = 0.1

julia> op(A, b)
Paving:
- tolerance ϵ = 0.1
- inner approx. of length 1188
- boundary approx. of length 821

IntervalLinearAlgebra.comparison_matrix — Method

comparison_matrix(A::Matrix{Interval})

Computes the comparison matrix ⟨A⟩ of the given matrix A according to the definition ⟨A⟩ii = mig(Aii) ⟨A⟩ij = -mag(Aij)

IntervalLinearAlgebra.enclose — Method

enclose(A::Matrix{Interval}, b::Vector{Interval})

Computes an initial enclosure Σ so that x ⊆ Σ, where x is the solution of the interval system Ax = b.

IntervalLinearAlgebra.interval_norm — Method

interval_norm(A::Matrix{Interval})

computes the infinity norm of interval matrix A.

IntervalLinearAlgebra.is_H_matrix — Method

is_H_matrix(A)

Tests whether an interval matrix A is an H-matrix, by testing that ⟨A⟩⁻¹e>0, where e=[1, 1, …, 1]ᵀ. Note that in practice it tests that an approximation of ⟨A⟩⁻¹e satisfies the condition. For more details see section 4.4 of [HOR19]

IntervalLinearAlgebra.is_M_matrix — Method

is_M_matrix(A)

Checks whether the interval matrix A is an M-matrix, that is a Z-matrix with non-negative inverse. For more details see section 4.2 of [HOR19]

IntervalLinearAlgebra.is_Z_matrix — Method

is_Z_matrix(A)

Checks whether the interval matrix A is a Z-matrix, that is whether Aᵢⱼ≤0 for all i≠j. For more details see section 4.2 of [HOR19]

IntervalLinearAlgebra.is_strictly_diagonally_dominant — Method

is_strictly_diagonally_dominant(A)

Checks whether an interval matrix A is stictly diagonally dominant, that is if mig(Aᵢᵢ) > ∑_(k ≠ i) mag(Aᵢₖ) for i=1,…,n. For more details see section 4.5 of [HOR19]

IntervalLinearAlgebra.is_strongly_regular — Method

is_strongly_regular(A)

Tests whether an interval matrix A is strongly regular. An interval matrix A is strongly regular if Ac⁻¹A is regular. For more details see section 4.6 of [HOR19]

IntervalLinearAlgebra.oettli_eq — Method

Returns the separator for the constraint |a_c ⋅x - b_c| - a_r ⋅|x| - b_r <= 0.

a and x must be vectors of the same length and b is scalar.

The absolue values in the equation are taken elementwise.

a_c and a_r are vectors containing midpoints and radii of the intervals in a. Similar b_c and b_r.

IntervalLinearAlgebra.oettli_lhs — Method

returns the unrolled expression for $|a ⋅x - b|$ $a$ and $x$ must be vectors of the same length and $b$ is scalar. The absolue value in the equation is taken elementwise.

IntervalLinearAlgebra.oettli_rhs — Method

returns the unrolled expression for $a ⋅|x| + b$ $a$ and $x$ must be vectors of the same length and $b$ is scalar. The absolue value in the equation is taken elementwise.

IntervalLinearAlgebra.rref — Method

rref(A)

computes the reduced row echelon form of the interval matrix A using maximum mignitude as pivoting strategy.

Parameters:

A : interval matrix