IntervalLinearAlgebra.eigenbox

IntervalLinearAlgebra.eigenbox — Method

eigenbox(A)

Returns an enclosure of all the eigenvalues of A. If A is symmetric, than the output is a real interval, otherwise it is a complex interval.

Algorithm

The algorithms used by the function are described in [HLA13].

Notes

The enclosure is not rigorous, meaning that the real eigenvalue problems solved internally utilize normal floating point computations.

Examples

julia> A = [0 -1 -1;2 -1.399.. -0.001 0;1 0.5 -1]
3×3 Matrix{Interval{Float64}}:
 [0, 0]  [-1, -1]                       [-1, -1]
 [2, 2]       [-1.39901, -0.000999999]    [0, 0]
 [1, 1]        [0.5, 0.5]               [-1, -1]

julia> eigenbox(A)
[-1.90679, 0.970154] + [-2.51903, 2.51903]im